Cleanup AddRadioEffect
Isolated the various FX being applied to separate Filter objects, that are now applied onto the input.

This should make it easier to add new post process, as well as give a better understanding of how they stack together.

Additionally, tweaked the Bandpass to be a single filter instead of multiple ones where the second would be essentially a noop.